+++ apps/phoenix_webcomponent/lib/phoenix_webcomponent/table.ex
@@ -6,34 +6,94 @@ defmodule Phoenix.WebComponent.Table do
 
   @doc """
   Generates a table.
-  ## Examples
-  ## Options
-  """
-  def wc_table(assigns) do
-    assigns =
-      assigns
-      |> assign_new(:id, fn -> false end)
-      |> assign_new(:class, fn -> "" end)
-      |> assign_new(:col, fn -> [] end)
-      |> assign_new(:rows, fn -> [] end)
 
+  ## Examples
+
+  ```heex
+  <.wc_table rows={[
+    %{
+      name: "Shmi Skywalker",
+      portrayal: "Pernilla August (Episodes I-II)"
+    },
+    %{
+      name: "Luke Skywalker",
+      portrayal: "Mark Hamill (Episodes IV-IX, The Mandalorian, The Book of Boba Fett)"}
+    ]}>
+    <:col let={r} label="Name">
+      <%= r.name %>
+    </:col>
+    <:col let={r} label="Portrayal">
+      <%= r.portrayal %>
+    </:col>
+  </.wc_table>
+  ```
+
+  """
+  @doc type: :component
+  attr(:id, :any,
+    default: false,
+    doc: """
+    html attribute id
+    """
+  )
+
+  attr(:class, :string,
+    default: "",
+    doc: """
+    html attribute class
+    """
+  )
+
+  attr(:rows, :list,
+    default: [],
+    doc: """
+    table data list
+    """
+  )
+
+  slot(:col,
+    required: false,
+    doc: """
+    render a column of table.
+
+    Example
+    ```heex
+    <:col let={r} label="Name">
+      <%= r.name %>
+    </:col>
+    ```
+    """
+  ) do
+    attr(:label, :string, doc: "table column title")
+  end
+
+  def wc_table(assigns) do
     ~H"""
-    <table id={@id} class={"table-fixed border-collapse border-spacing-0 #{@class}"}>
+    <table
+      id={@id}
+      class={[
+        "table-fixed border-collapse border-spacing-0",
+        @class,
+      ]}
+    >
       <thead>
         <tr class="bg-slate-100 h-[34px]">
-          <%= for col <- @col do %>
-            <th class=" px-4 py-2 slate-700 text-left font-medium"><%= col.label %></th>
-          <% end %>
+          <th
+            :for={col <- @col}
+            class="px-4 py-2 slate-700 text-left font-medium"
+          ><%= col.label %></th>
         </tr>
       </thead>
       <tbody>
-      <%= for row <- @rows do %>
-        <tr class="bg-slate-50 even:bg-white h-[40px]">
-          <%= for col <- @col do %>
-            <td class=" px-4 py-2"><%= render_slot(col, row) %></td>
-          <% end %>
+        <tr
+          :for={row <- @rows}
+          class={"bg-slate-50 even:bg-white h-[40px]"}
+        >
+          <td
+            :for={col <- @col}
+            class="px-4 py-2"
+          ><%= render_slot(col, row) %></td>
         </tr>
-      <% end %>
       </tbody>
     </table>
     """