+++ apps/phx_wc_storybook_web/lib/phx_wc_storybook_web/router.ex
@@ -15,14 +15,18 @@ defmodule PhxWCStorybookWeb.Router do
     plug :accepts, ["json"]
   end
 
+  scope "/" do
+    storybook_assets()
+  end
+
   scope "/", PhxWCStorybookWeb do
     pipe_through :browser
 
     get "/", PageController, :index
-  end
 
-  live_storybook("/storybook",
-    otp_app: :phx_wc_storybook_web,
-    backend_module: PhxWCStorybookWeb.Storybook
-  )
+    get "/mdi", PageController, :mdi
+    get "/bsi", PageController, :bsi
+
+    live_storybook("/storybook", backend_module: PhxWCStorybookWeb.Storybook)
+  end
 end